Initial round of refactoring
